Make ESCAPE-DISPLAY-NAME and ESCAPE-LOCAL-PART unicode aware
Before this commit, characters outside of the ASCII range would cause
the component to always be surrounded by quotes, and characters outside
of the ASCII range to be preceded by a backslash.

Non-ASCII characters are now left alone. This commit makes the handling
of such characters better aligned with the parser's support in case,
*ALLOW-UNICODE* is true, where such special treatment of non-ASCII characters
is not required.